Update Quick Start architecture diagram.pptx
I based this file on the master deck. Pasting items from the master deck used to change the font to Calibri and change certain line colors and arrows to look different (unless people thought to say "Keep source formatting" with each paste). Now, we don't have to think about this. The source formatting is this file's formatting now match.